// PinaFront/src/app/services/models/artista.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Opera } from '../models/opera';
export interface Artista {
  cognome?: string;
  dataDiMorte?: string;
  dataDiNascita?: string;
  id?: number;
  immagine?: string;
  luogoDiNascita?: string;
  name?: string;
  opere?: Array<Opera>;
}