// PinaFront/src/app/services/models/opera-front.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ArtistaFrontImg } from '../models/artista-front-img';
export interface OperaFront {
  anno?: number;
  artista?: ArtistaFrontImg;
  id?: number;
  img?: string;
  posizione?: number;
  tecnica?: string;
  title?: string;
}