// PinaFront/src/app/services/models/area-front.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { OperaFront } from '../models/opera-front';
import { UserFront } from '../models/user-front';
export interface AreaFront {
  curatore?: UserFront;
  id?: number;
  items?: Array<OperaFront>;
  name?: string;
}